Add HTML framework, calculation JS, basic CSS
Added a grid of buttons and a screen to represent a calculator.
Added a div with boilerplate text to the right of the grid.
Added flex to to calculator and body.
Added sum, subtract, divide, and multiply functions.
Added an operate function to accept an operation value and pass other args as parts of the calculation.
Added Instructions from TOP to readme.me